Update README.md
So this game is supposed to be similar to the prototype developed in the seminar: as in, the player shoots stuff that falls down and is able to collect stuff.
Its just that there'S enough Corona in my life already so I wanted to change the setting, preferrably to something where I may be able to find prettier background pictures.
Also, I wanted the player to be allowed to go/fly/swim higher than just the edge of the game view screen, so I wanted an environment where that made sense aesthetically. I figured I could do something underwater, because then when the player can reach the top of the ocean and it can start with darker images and get lighter when it gets closer towards the top/the sun; so it signals the end. The player ended up being a Kraken 
( - and I guess also because subconciously I liked that start screen from GitKraken, but I didnt notice up until I liked the rest of it so much I didnt want to change it -)
because I figured that they have that ink-thing so it makes sense that they shoot at something. They end up shooting at plastic/other rubbish clogging the oceans because, ya know, the environment. 

In the game the going higher then was supposed to include reaching upper levels, but since I didnt want it to be that you can just straight fly to the end, there ended up being keys one needs to collect to be allowed to pass. Other collectibles are supposed to include: getting faster, getting coins for the highscore, getting more lives/protection, collecting items that allow you to shoot more powerfully.

If there is time left, I'd like to include fake items that look like the power-ups but instead of enhance, block the players powers/ability instead (so go slower, shoot slower, lose score points).