Update README.md
In the second commit, the game now has most of its basic functions. You collect PowerUps that help you, and keys to reach the next Level, you get a score and a message each for a game either lost or won. Most of the thing tryin to harm the player change each level (except for the jellyfish because I thought they were really cool looking).
There is an enemy at the top that needs to be defeated to win the game so that it doesnt go on endlessly, but one can still try to beat the score. 

What needs to be done next:
Obviously the game needs to look prettier, so I'll work on that first. I assume that'll take some time. I need to correct small issues like one PowerUp either rotating wildly or moving up - but not both at the same time. 

If there is time left, Id like to include more PowerUps, like a shield that isnt connected to the player but appears locally, so the player can hide underneath it; and it gets destroyed either if hit or after a certain time. Also I want PowerDowns (slower movement, shooting) and at least partially connect them to PowerUps that are kinda too strong (like the wave: maybe after the player used it, the player can't shoot for x seconds). Also, a StartScreen/Button and one that explains the game míght look cool, and something like a personal scoreboard, if possible.